+++ src/digitalWriteFast.h
@@ -439,7 +439,7 @@
 
 
 #ifndef digitalWriteFast
-#if (defined(__AVR__) || defined(ARDUINO_ARCH_AVR))
+#if (defined(__AVR__) || defined(ARDUINO_ARCH_AVR)) && defined(__digitalPinToPortReg)
 #define digitalWriteFast(P, V) \
 if (__builtin_constant_p(P)) { \
   BIT_WRITE(*__digitalPinToPortReg(P), __digitalPinToBit(P), (V)); \
@@ -453,7 +453,7 @@ if (__builtin_constant_p(P)) { \
 
 
 #ifndef pinModeFast
-#if (defined(__AVR__) || defined(ARDUINO_ARCH_AVR))
+#if (defined(__AVR__) || defined(ARDUINO_ARCH_AVR)) && defined(__digitalPinToPortReg)
 #define pinModeFast(P, V) \
 if (__builtin_constant_p(P) && __builtin_constant_p(V)) { \
   if (V == INPUT_PULLUP) {\