+++ src/ir_LG.cpp
@@ -151,9 +151,9 @@ void IRsend::sendLGRaw(uint32_t aRawData, uint_fast8_t aNumberOfRepeats, bool aI
     for (uint_fast8_t i = 0; i < aNumberOfRepeats; ++i) {
         // send repeat in a 110 ms raster
         if (i == 0) {
-            delay(LG_REPEAT_SPACE / 1000);
+            delay(LG_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
         } else {
-            delay((LG_REPEAT_PERIOD - LG_REPEAT_DURATION) / 1000);
+            delay((LG_REPEAT_PERIOD - LG_REPEAT_DURATION) / MICROS_IN_ONE_MILLI);
         }
         // send repeat
         sendLGRepeat(aUseLG2Protocol);