+++ src/ir_RC5_RC6.cpp
@@ -109,7 +109,7 @@ void IRsend::sendRC5(uint8_t aAddress, uint8_t aCommand, uint_fast8_t aNumberOfR
         // skip last delay!
         if (tNumberOfCommands > 0) {
             // send repeated command in a fixed raster
-            delay(RC5_REPEAT_SPACE / 1000);
+            delay(RC5_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
         }
     }
 }
@@ -335,7 +335,7 @@ void IRsend::sendRC6(uint8_t aAddress, uint8_t aCommand, uint_fast8_t aNumberOfR
         // skip last delay!
         if (tNumberOfCommands > 0) {
             // send repeated command in a fixed raster
-            delay(RC6_REPEAT_SPACE / 1000);
+            delay(RC6_REPEAT_SPACE / MICROS_IN_ONE_MILLI);
         }
     }
 }